refactor: Remove unused GitHub Actions workflow for publishing to GitHub Packages
The commit removes the unused GitHub Actions workflow file ".github/workflows/npm-publish-github-packages.yml" that was responsible for running tests and publishing a package to GitHub Packages when a release was created. This workflow is no longer needed and has been replaced by a new workflow for publishing to npm.

The changes include:
- Deleting the file ".github/workflows/npm-publish-github-packages.yml"